% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/play.R
\name{play}
\alias{play}
\title{R Slot Machine}
\usage{
play(
  n = 1,
  betsize = 1,
  cash = 10,
  prob = c(0.03, 0.03, 0.06, 0.1, 0.25, 0.01, 0.52)
)
}
\arguments{
\item{n}{number of plays}

\item{betsize}{default is 1.}

\item{cash}{amount of starting cash}

\item{prob}{probabilities for getting diamonds (DD), 7s, bars (B,BB, BBB for single, double and triples),Cherries (C) and Zeros (0). Change the probabilities to change the odds of winning!}
}
\description{
A slot machine simulator in R. Based on the example outlined in \href{https://rstudio-education.github.io/hopr/project-3-slot-machine.html}{Hands on in R Programming}. The default configuration seems to have multiple plays (`n>1`) lead to being bankrupt. However it appears that setting the `betsize` larger leads to more profitable returns when using the default probabilities.
}
\examples{
library(rslots)
res<- play(n=10, cash=100)
res
}
